+++ src/numerical/index.ts
@@ -8,3 +8,4 @@ export { hasEvenLength } from './hasEvenLength.js';
 export { median } from './median.js';
 export { mode } from './mode.js';
 export { range } from './range.js';
+export { variance } from './variance.js';